+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.1.0
 
 Initial revision.
+
+## 0.1.1
+
+- Added support for circular references.
+- Built a full test suite, which now passes!